thumbnails (#26)
* add image-size and @types/sharp deps

* draft ContentStorage

* add bullmq dependency

* create and listen to jobs (bullmq)

* listen to jobs

* fix default blocked beneficiary

* setDevice find before create

* new upload methods

* new env vars + refactor config

* new upload community picture endpoint

* remove txCreationObj; add logo

* add swagger

* draft ContentStorage

* delete methods

* refactor content storage class

* change add story endpoint

* ignore storage redis on tests

* change thumbnail generate name

* app media content and thumbnail tables

* refactor storage class to better suppt thumbnails

* fix typo

* adjustments to vairable names

* fix association

* include thumbnails in community result

* update story to use content media and thumbnails

* refactor list light + endpoint validation

* remove unused

* edit community

* refactor deleteStory

* prevent some db migrations to run during tests

* fix test

* issue fix + remove some duplicated code

* refactor content storage class

* merge duplications

* empty method

* fix merge issue

* remove community logo

* fix minimal issues

+++ src/database/migrations/z1611675848-reach-in-out.js
@@ -2,6 +2,9 @@
 
 module.exports = {
     async up(queryInterface, Sequelize) {
+        if (process.env.NODE_ENV === 'test') {
+            return;
+        }
         await queryInterface.sequelize.query('delete from reachedaddress');
 
         let query;
