+++ src/database/migrations/z1612567510-delete-0-ssi.js
@@ -1,7 +1,10 @@
 'use strict';
 
 module.exports = {
-    up(queryInterface, Sequelize) {
+    async up(queryInterface, Sequelize) {
+        if (process.env.NODE_ENV === 'test') {
+            return;
+        }
         return queryInterface.sequelize.query(
             'delete from ubi_community_daily_metrics where ssi = 0'
         );
